# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coral Companion API Documentation",
        "description": "Documentation for the API used in the Coral Companion app",
        "contact": {
            "name": "Paula Rumeu",
            "email": "paururo08@hotmail.com"
        },
        "license": {
            "name": "Creative Commons Attribution Share Alike 4.0 International",
            "url": "https://creativecommons.org/licenses/by-sa/4.0/"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://github.com/salem404/coral-companion/raw/master/frontend/src/assets/img/logo-color.svg"
        }
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Localhost"
        },
        {
            "url": "https://coral-companion.railway.com/api",
            "description": "Railway"
        }
    ],
    "paths": {
...
        "/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get all tasks",
                "description": "Returns all tasks",
                "operationId": "b132a5f14cb1c055f33ef7d94af79438",
                "responses": {
                    "404": {
                        "description": "Not found: Tasks don't exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No tasks found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Create a task",
                "description": "Create a new task",
                "operationId": "4dbf3e760f46296529098c64a16b6549",
                "responses": {
                    "201": {
                        "description": "Task created successfully"
                    },
                    "401": {
                        "description": "You are not authorized to create a task"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get a task by id",
                "description": "Returns a task",
                "operationId": "42e07aa4bda2700f1c22a049e01cd7a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Task"
                ],
                "summary": "Update a task",
                "description": "Update a task",
                "operationId": "caa557a9905d8dc7f6b329db3d4b6a98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task updated successfully"
                    },
                    "401": {
                        "description": "You are not authorized to update this task"
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Delete a task",
                "description": "Delete a task",
                "operationId": "6f5697c2a7171f21f3455412a4ea996d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task deleted successfully"
                    },
                    "401": {
                        "description": "You are not authorized to delete this task"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}